Consolidate test files
