+++ package.json
@@ -40,20 +40,20 @@
     "pretest:contracts:extensions": "pnpm run stop:blockchain:client",
     "test:contracts": "pnpm run start:blockchain:client && npx hardhat test ./test/contracts-network/* ./test/packages/* --network development",
     "test:contracts:extensions": "pnpm run start:blockchain:client && npx hardhat test ./test/extensions/* --network development",
-    "test:contracts:upgrade:parity": "pnpm run clean:test:contracts | pnpm run start:blockchain:client parity & pnpm run generate:test:contracts && npx hardhat test ./test-upgrade/* --network integration",
-    "test:contracts:upgrade:ganache": "pnpm run clean:test:contracts | pnpm run start:blockchain:client ganache & pnpm run generate:test:contracts && npx hardhat test ./test-upgrade/* --network development",
-    "test:contracts:gasCosts": "pnpm run start:blockchain:client && npx hardhat test test-gas-costs/gasCosts.js --network development",
+    "test:contracts:upgrade:parity": "pnpm run clean:test:contracts | pnpm run start:blockchain:client parity & pnpm run generate:test:contracts && npx hardhat test ./test/upgrade/* --network integration",
+    "test:contracts:upgrade:ganache": "pnpm run clean:test:contracts | pnpm run start:blockchain:client ganache & pnpm run generate:test:contracts && npx hardhat test ./test/upgrade/* --network development",
+    "test:contracts:gas": "pnpm run start:blockchain:client && npx hardhat test test/misc/gas-costs.js --network development",
     "test:contracts:patricia": "pnpm run start:blockchain:client ganache && npx hardhat test packages/reputation-miner/patricia-test.js --network integration",
     "test:contracts:bridging:1": "CHAIN_ID_1=265669100 CHAIN_ID_2=265669101 pnpm run start:blockchain:client:both && HARDHAT_FOREIGN=false MINING_CHAIN_ID=265669100 npx hardhat test ./test/cross-chain/* --network development",
     "test:contracts:bridging:2": "CHAIN_ID_1=265669101 CHAIN_ID_2=265669100 pnpm run start:blockchain:client:both && HARDHAT_FOREIGN=true MINING_CHAIN_ID=265669100 CHAIN_ID=265669101 npx hardhat test ./test/cross-chain/* --network development",
-    "test:contracts:chainid": "pnpm run start:blockchain:client && npx hardhat test ./test-chainid/*",
-    "test:contracts:e2e": "pnpm run start:blockchain:client && npx hardhat test test-system/end-to-end.js --network development",
+    "test:contracts:chainid": "pnpm run start:blockchain:client && npx hardhat test ./test/misc/chainid.js",
+    "test:contracts:e2e": "pnpm run start:blockchain:client && npx hardhat test test/misc/end-to-end.js --network development",
     "test:contracts:coverage": "npx hardhat coverage --temp build-coverage --testfiles './test/contracts-network/*'",
     "test:contracts:extensions:coverage": "npx hardhat compile && NODE_OPTIONS='--max-old-space-size=6144' npx hardhat coverage --solcoverjs ./.solcover.extensions.js --temp build-coverage --testfiles './test/extensions/*'",
     "test:contracts:bridging:1:coverage": "CHAIN_ID=265669101 pnpm run start:blockchain:client:2 && HARDHAT_FOREIGN=false MINING_CHAIN_ID=265669100 npx hardhat --show-stack-traces coverage --solcoverjs ./.solcover.crosschain.js --temp build-coverage --testfiles './test/cross-chain/**/*'",
     "test:contracts:bridging:2:coverage": "CHAIN_ID=265669100 pnpm run start:blockchain:client:2 && HARDHAT_FOREIGN=true MINING_CHAIN_ID=265669100 CHAIN_ID=265669101 npx hardhat coverage --solcoverjs ./.solcover.crosschain.js --temp build-coverage --testfiles './test/cross-chain/**/*'",
-    "test:contracts:chainid:coverage": "npx hardhat coverage --solcoverjs ./.solcover.chainid.js --temp build-coverage --testfiles './test-chainid/**/*'",
-    "test:contracts:upgrade:coverage": "pnpm run generate:test:contracts && npx hardhat coverage --solcoverjs ./.solcover.upgrade.js --temp build-coverage --testfiles './test-upgrade/*'",
+    "test:contracts:chainid:coverage": "npx hardhat coverage --solcoverjs ./.solcover.chainid.js --temp build-coverage --testfiles './test/misc/chainid.js'",
+    "test:contracts:upgrade:coverage": "pnpm run generate:test:contracts && npx hardhat coverage --solcoverjs ./.solcover.upgrade.js --temp build-coverage --testfiles './test/upgrade/*'",
     "test:reputation:1": "pnpm run start:blockchain:client && npx hardhat test ./test/reputation-system/*.js --network development",
     "test:reputation:2": "pnpm run start:blockchain:client && npx hardhat test ./test/reputation-system/reputation-mining-client/* --network development",
     "test:reputation:1:anotherChain": "CHAIN_ID=101010101 pnpm run start:blockchain:client && MINING_CHAIN_ID=101010101 CHAIN_ID=101010101 npx hardhat test ./test/reputation-system/*.js --network development",
@@ -64,7 +64,7 @@
     "posttest:contracts:extensions": "pnpm run stop:blockchain:client",
     "posttest:contracts:upgrade:parity": "pnpm run clean:test:contracts | pnpm run stop:blockchain:client",
     "posttest:contracts:upgrade:ganache": "pnpm run clean:test:contracts | pnpm run stop:blockchain:client",
-    "posttest:contracts:gasCosts": "pnpm run stop:blockchain:client",
+    "posttest:contracts:gas": "pnpm run stop:blockchain:client",
     "posttest:contracts:patricia": "pnpm run stop:blockchain:client",
     "viz:bootstrap": "pnpm run start:blockchain:client && HARDHAT_NETWORK=development npx hardhat run scripts/viz-bootstrap.js",
     "kyc:bootstrap": "pnpm run start:blockchain:client && HARDHAT_NETWORK=development npx hardhat run scripts/kyc-bootstrap.js"
